Add functionality to locations migration

+++ database/migrations/2015_07_14_081220_create_assetSubclass_table.php
@@ -34,6 +34,9 @@ class CreateAssetSubclassTable extends Migration
      */
     public function down()
     {
-        Schema::drop('asset_subclasses');
+        if(Schema::hasTable('asset_subclasses'))
+        {
+           Schema::drop('asset_subclasses');
+        }
     }
 }
